moved the food counter to home.js
