+++ src/Pages/Home.js
@@ -2,6 +2,8 @@ import React from 'react'
 import '../Components/styles/Home.css'
 import Ray from '../Components/Ray'
 import { BrowserRouter as Router, Switch, Route, Link } from 'react-router-dom'
+import 'firebase/database'
+import * as firebase from 'firebase/app'
 
 import Assignment from './Assignment'
 import Food from './Food'
@@ -12,14 +14,65 @@ import Bar from '../Components/Bar'
 class Home extends React.Component {
 	constructor(props) {
 		super(props)
-		
+		this.state = {
+			availableFood: 0,
+			seconds: 3600
+		}
+		//refills food over time
+		this.timer = 0
+		this.startTimer = this.startTimer.bind(this)
+		this.countDown = this.countDown.bind(this)
 	}
 
 	componentDidMount() {
 		fetch('http://192.168.86.28:8080/uwu').then((res) => {})
+		var firebaseConfig = {
+			apiKey: 'AIzaSyAImG5Vk9cS8Yi_UUNX9gwO-4_b1z2KAR0',
+			authDomain: 'rayside-94e8d.firebaseapp.com',
+			databaseURL: 'https://rayside-94e8d.firebaseio.com',
+			projectId: 'rayside-94e8d',
+			storageBucket: 'rayside-94e8d.appspot.com',
+			messagingSenderId: '819405678971',
+			appId: '1:819405678971:web:74554bcb338cafdb07b5de',
+			measurementId: 'G-4JECV8ZB79',
+		}
+
+		if (!firebase.apps.length) {
+			firebase.initializeApp(firebaseConfig)
+		}
+		var database = firebase.database()
+		database.ref('food/food').on('value', (snapshot) => {
+			this.setState({ availableFood: snapshot.val() });
+			if (snapshot.val() < 10) {
+				this.startTimer();
+			}
+		})
 	}
+
 	componentWillUnmount() {
 		window.clearInterval()
+	}
+
+	startTimer() {
+		if (this.timer == 0 && this.state.seconds > 0) {
+			this.timer = setInterval(this.countDown, 1000);
+		}
+	}
+
+	countDown() {
+		let current_seconds = this.state.seconds-1;
+		this.setState({
+			seconds: current_seconds
+		});
+		if (current_seconds == 0) {
+			clearInterval(this.timer);
+			this.timer = 0;
+			this.setState({
+				seconds: 5
+			})
+			var database = firebase.database()
+			database.ref('food/food').set(this.state.availableFood+1)
+		}
 	}
 
 	render() {
